remove css support prototype
