+++ packages/loader/web-code-loader/src/index.ts
@@ -6,3 +6,4 @@ export * from "./semVerCdnCodeResolver";
 export * from "./utils";
 export * from "./webLoader";
 export * from "./allowList";
+export * from "./browserPackage";
